move output computations to positron notebook editor
